Struct solana_test_validator::TestValidatorGenesis

pub struct TestValidatorGenesis {

    pub(crate) fee_rate_governor: FeeRateGovernor,
    pub(crate) ledger_path: Option<PathBuf>,
    pub(crate) tower_storage: Option<Arc<dyn TowerStorage>>,
    pub rent: Rent,
    pub(crate) rpc_config: JsonRpcConfig,
    pub(crate) pubsub_config: PubSubConfig,
    pub(crate) rpc_ports: Option<(u16, u16)>,
    pub(crate) warp_slot: Option<Slot>,
    pub(crate) no_bpf_jit: bool,
    pub(crate) accounts: HashMap<Pubkey, AccountSharedData>,
    pub(crate) programs: Vec<ProgramInfo>,
    pub(crate) upgradeable_programs: Vec<UpgradeableProgramInfo>,
    pub(crate) ticks_per_slot: Option<u64>,
    pub(crate) epoch_schedule: Option<EpochSchedule>,
    pub(crate) node_config: TestValidatorNodeConfig,
    pub validator_exit: Arc<RwLock<Exit>>,
    pub start_progress: Arc<RwLock<ValidatorStartProgress>>,
    pub authorized_voter_keypairs: Arc<RwLock<Vec<Arc<Keypair>>>>,
    pub staked_nodes_overrides: Arc<RwLock<HashMap<Pubkey, u64>>>,
    pub max_ledger_shreds: Option<u64>,
    pub max_genesis_archive_unpacked_size: Option<u64>,
    pub geyser_plugin_config_files: Option<Vec<PathBuf>>,
    pub(crate) deactivate_feature_set: HashSet<Pubkey>,
    pub(crate) compute_unit_limit: Option<u64>,
    pub log_messages_bytes_limit: Option<usize>,
    pub transaction_account_lock_limit: Option<usize>,
    pub tpu_enable_udp: bool,
    pub geyser_plugin_manager: Arc<RwLock<GeyserPluginManager>>,
    pub(crate) admin_rpc_service_post_init: Arc<RwLock<Option<AdminRpcRequestMetadataPostInit>>>,

}

Fields
fee_rate_governor: FeeRateGovernorledger_path: Option<PathBuf>tower_storage: Option<Arc<dyn TowerStorage>>rent: Rentrpc_config: JsonRpcConfigpubsub_config: PubSubConfigrpc_ports: Option<(u16, u16)>warp_slot: Option<Slot>no_bpf_jit: boolaccounts: HashMap<Pubkey, AccountSharedData>programs: Vec<ProgramInfo>upgradeable_programs: Vec<UpgradeableProgramInfo>ticks_per_slot: Option<u64>epoch_schedule: Option<EpochSchedule>node_config: TestValidatorNodeConfigvalidator_exit: Arc<RwLock<Exit>>start_progress: Arc<RwLock<ValidatorStartProgress>>authorized_voter_keypairs: Arc<RwLock<Vec<Arc<Keypair>>>>staked_nodes_overrides: Arc<RwLock<HashMap<Pubkey, u64>>>max_ledger_shreds: Option<u64>max_genesis_archive_unpacked_size: Option<u64>geyser_plugin_config_files: Option<Vec<PathBuf>>deactivate_feature_set: HashSet<Pubkey>compute_unit_limit: Option<u64>log_messages_bytes_limit: Option<usize>transaction_account_lock_limit: Option<usize>tpu_enable_udp: boolgeyser_plugin_manager: Arc<RwLock<GeyserPluginManager>>admin_rpc_service_post_init: Arc<RwLock<Option<AdminRpcRequestMetadataPostInit>>>
Implementations

impl TestValidatorGenesis

pub fn deactivate_features(&mut self, deactivate_list: &[Pubkey]) -> &mut Self

Adds features to deactivate to a set, eliminating redundancies during initialize_ledger, if member of the set is not a Feature it will be silently ignored

pub fn ledger_path<P: Into<PathBuf>>(&mut self, ledger_path: P) -> &mut Self

pub fn tower_storage( &mut self, tower_storage: Arc<dyn TowerStorage> ) -> &mut Self

pub fn ledger_exists(ledger_path: &Path) -> bool

Check if a given TestValidator ledger has already been initialized

pub fn tpu_enable_udp(&mut self, tpu_enable_udp: bool) -> &mut Self

pub fn fee_rate_governor( &mut self, fee_rate_governor: FeeRateGovernor ) -> &mut Self

pub fn ticks_per_slot(&mut self, ticks_per_slot: u64) -> &mut Self

pub fn epoch_schedule(&mut self, epoch_schedule: EpochSchedule) -> &mut Self

pub fn rent(&mut self, rent: Rent) -> &mut Self

pub fn rpc_config(&mut self, rpc_config: JsonRpcConfig) -> &mut Self

pub fn pubsub_config(&mut self, pubsub_config: PubSubConfig) -> &mut Self

pub fn rpc_port(&mut self, rpc_port: u16) -> &mut Self

pub fn faucet_addr(&mut self, faucet_addr: Option<SocketAddr>) -> &mut Self

pub fn warp_slot(&mut self, warp_slot: Slot) -> &mut Self

pub fn bpf_jit(&mut self, bpf_jit: bool) -> &mut Self

pub fn gossip_host(&mut self, gossip_host: IpAddr) -> &mut Self

pub fn gossip_port(&mut self, gossip_port: u16) -> &mut Self

pub fn port_range(&mut self, port_range: PortRange) -> &mut Self

pub fn bind_ip_addr(&mut self, bind_ip_addr: IpAddr) -> &mut Self

pub fn compute_unit_limit(&mut self, compute_unit_limit: u64) -> &mut Self

pub fn max_compute_units(&mut self, max_compute_units: u64) -> &mut Self

๐Deprecated: Please use compute_unit_limit instead

pub fn add_account( &mut self, address: Pubkey, account: AccountSharedData ) -> &mut Self

Add an account to the test environment

pub fn add_accounts<T>(&mut self, accounts: T) -> &mut Selfwhere T: IntoIterator<Item = (Pubkey, AccountSharedData)>,

pub fn clone_accounts<T>( &mut self, addresses: T, rpc_client: &RpcClient, skip_missing: bool ) -> Result<&mut Self, String>where T: IntoIterator<Item = Pubkey>,

pub fn clone_upgradeable_programs<T>( &mut self, addresses: T, rpc_client: &RpcClient ) -> Result<&mut Self, String>where T: IntoIterator<Item = Pubkey>,

pub fn add_accounts_from_json_files( &mut self, accounts: &[AccountInfo<'_>] ) -> Result<&mut Self, String>

pub fn add_accounts_from_directories<T, P>( &mut self, dirs: T ) -> Result<&mut Self, String>where T: IntoIterator<Item = P>, P: AsRef<Path> + Display,

pub fn add_account_with_file_data( &mut self, address: Pubkey, lamports: u64, owner: Pubkey, filename: &str ) -> &mut Self

Add an account to the test environment with the account data in the provided filename

pub fn add_account_with_base64_data( &mut self, address: Pubkey, lamports: u64, owner: Pubkey, data_base64: &str ) -> &mut Self

Add an account to the test environment with the account data in the provided as a base 64 string

pub fn add_program( &mut self, program_name: &str, program_id: Pubkey ) -> &mut Self

Add a SBF program to the test environment.

program_name will also used to locate the SBF shared object in the current or fixtures directory.

pub fn add_programs_with_path(&mut self, programs: &[ProgramInfo]) -> &mut Self

๐Deprecated since 1.16.0: Please use add_upgradeable_programs_with_path() instead

Add a list of programs to the test environment.

pub fn add_upgradeable_programs_with_path( &mut self, programs: &[UpgradeableProgramInfo] ) -> &mut Self

Add a list of upgradeable programs to the test environment.

pub fn start_with_mint_address( &self, mint_address: Pubkey, socket_addr_space: SocketAddrSpace ) -> Result<TestValidator, Box<dyn Error>>

Start a test validator with the address of the mint account that will receive tokens created at genesis.

pub fn start_with_mint_address_and_geyser_plugin_rpc( &self, mint_address: Pubkey, socket_addr_space: SocketAddrSpace, rpc_to_plugin_manager_receiver: Option<Receiver<GeyserPluginManagerRequest>> ) -> Result<TestValidator, Box<dyn Error>>

Start a test validator with the address of the mint account that will receive tokens created at genesis. Augments admin rpc service with dynamic geyser plugin manager if the geyser plugin service is enabled at startup.

pub fn start(&self) -> (TestValidator, Keypair)

Start a test validator

Returns a new TestValidator as well as the keypair for the mint account that will receive tokens created at genesis.

This function panics on initialization failure.

pub fn start_with_socket_addr_space( &self, socket_addr_space: SocketAddrSpace ) -> (TestValidator, Keypair)

Start a test validator with the given SocketAddrSpace

Returns a new TestValidator as well as the keypair for the mint account that will receive tokens created at genesis.

This function panics on initialization failure.

pub async fn start_async(&self) -> (TestValidator, Keypair)

pub async fn start_async_with_socket_addr_space( &self, socket_addr_space: SocketAddrSpace ) -> (TestValidator, Keypair)